+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - showing rule files in context files tree in Agent mode
 - fixed issue with caching when using OpenRouter provider and Anthropic models
 - watch files functionality now correctly only triggers once per project
+- updated close button text in Changelog dialog
 
 ## [0.41.0]